requirements.txt
alembic==1.18.1
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
Flask==3.1.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
greenlet==3.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.3
SQLAlchemy==2.0.45
typing_extensions==4.15.0
Werkzeug==3.1.5
Flask-Babel==4.0.0
pytest==8.3.4
python-dotenv==1.0.1
